refactor: WeatherConditionの切り分け
